requirements.txt
numpy
matplotlib
tqdm
pandas
pyserial
scikit-learn
pykan
sympy
torch
tueplots
scipy